+++ spec/cli/dummy/transpile_me.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class A
   def self.transform(val)
     val in {status:}