+++ spec/integration/fixtures/array_in_hash_pattern.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 case JSON.parse(ARGV[0], symbolize_names: true)
 in {name: "Alice", children: [{name: "Bob", age: age}]}
   p "Bob age is #{age}"