+++ spec/integration/fixtures/rubocop/numbered_params.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+proc { _1 + _2 }
+
+[1, 2].map do
+  _1 * 2
+end
+
+-> { _1 }